Fix bugs during training with cxr-bert transformer
Changed the transformer to use the entire cxr-bert model instead of just
the first part. Simplified the encode_text function to use the
get_projected_text_embeddings function instead of using a direct forward
call of the model. Also added back in the text_projection because using
the full model changes the transformer width so it needs to be
projected.