Add optional argument to train with CheXzero
Before, we had commented out the CheXzero model code. This change
refactors the code and adds an additional optional argument so that the
new model and the CheXzero model can both be trained. The --use_chexzero
flag is what you need to add to the run_train command to train the
CheXzero model. Otherwise, the new model will be trained.