Add locking functionality for model training (#9)
* Create params list for optimizer

Before we had been generating the parameters directly in the line where
the optimizer is defined. This change creates a params list first that
is passed into the optimizer.

* Rename use_chexzero flag to use_chexzero_text

Before the optional argument to run training with the chexzero text
transformer was use_chexzero. This change renames it to
use_chexzero_text to be more specific.

* Add ability to lock parts of model in training

Before you had to manually change the code to lock different parts of
the model during training. This change introduces the lock_text and
lock_vision flags that can be used with the run_train command. These
lock the corresponding stacks of the model during training.

* Add documentation for locking parts of model

This change adds documentation in the README file to help create a
given experimental setup that would be used while training a model. It
includes info on locking text stack, locking vision stack, and using the
chexzero text transformer.

* Create params list for optimizer

Before we had been generating the parameters directly in the line where
the optimizer is defined. This change creates a params list first that
is passed into the optimizer.

* Rename use_chexzero flag to use_chexzero_text

Before the optional argument to run training with the chexzero text
transformer was use_chexzero. This change renames it to
use_chexzero_text to be more specific.

* Add ability to lock parts of model in training

Before you had to manually change the code to lock different parts of
the model during training. This change introduces the lock_text and
lock_vision flags that can be used with the run_train command. These
lock the corresponding stacks of the model during training.

* Add documentation for locking parts of model

This change adds documentation in the README file to help create a
given experimental setup that would be used while training a model. It
includes info on locking text stack, locking vision stack, and using the
chexzero text transformer.

* Make the text_projection always get trained

Before the text_projection was only being trained when we were not
locking the text stack. This change makes sure we are always training
the text projection. The change also improves code maintanability by
using list comprehension at the end for the dictionary structure.

Co-authored-by: aakamishra <aakamishra1@gmail.com>