Exper docs (#13)
* Improve experiment documentation

This change updates the documentation to reflect the new use_cxrbert
flag instead of the previous use_chexzero_text flag. It also adds full
commands for the experiments to the documentation.

* Add documentation for vitmae experiments

This change adds documentation to the README file on how to run training
and experiments for the ViTMAE image stack.