Add some support for a few HuggingFace BERT models
Currently, we only support CXRBERT. This will add support for BlueBERT
and ClinicalBERT. In order for the models to be equivalent, we are
removing the MLP from CXRBERT and using a common simple linear projection
head trained from scratch for all models. The code is not fully functional
yet as we need to also add support a HuggingFace BERT key argument.