Image models (#19)
* Add Placeholder for uL Branch

* generic hugging face transformer

* Add some support for a few HuggingFace BERT models

Currently, we only support CXRBERT. This will add support for BlueBERT
and ClinicalBERT. In order for the models to be equivalent, we are
removing the MLP from CXRBERT and using a common simple linear projection
head trained from scratch for all models. The code is not fully functional
yet as we need to also add support a HuggingFace BERT key argument.

* Add HuggingFace BERT key support

The pipeline now supports three different pretrained BERT models from
Huggingface. A handful of bugs were also fixed.

* Add support for mococxr/medaug via BioViL classes

Co-authored-by: Christy Jestin <cjj9000@gmail.com>

* Add pretrained model dir to gitignore

* Add training code

* Add training parsing code

* Update arguments for run train

* Adjust learning rate

* Add flush

* Add wandb support

The pipeline now supports logging and saving of model checkpoints.

* Fix wandb config saving

Currently, the config isn't actually being updated or saved.

* Use custom transforms for BioViL

* Fix image size for biovil to also be 224

* Changes to vision projection for CLIP

* Fix zero shot for new model types

Co-authored-by: aakamishra <aakamishra1@gmail.com>
Co-authored-by: Rajat Mittal <rajatmittal@college.harvard.edu>
Co-authored-by: Christy Jestin <cjj9000@gmail.com>
Co-authored-by: christyjestin <christyjestin@college.harvard.edu>