VinDr-CXR support and evaluation (#20)
* Add Placeholder for uL Branch

* Add script which computes biovil chexpert AUC

* Move logic into functions

Previously code was structured as a notebook. This turns run_biovil into
runnable script that supports either generating the results or reading
the predictions from a csv and evaluating the AUC curves on them.

* Add predictions for biovil with biovil prompts

* Add new function definition with simple logic

* Load in biovil image inference engine

* Ignore cxr_data

* Add support for loading images for biovision

* Integrate BioViL Vision Stack into CheXzero Code

Added a flag called use_biovision and added in logic to load the BioViL
ResNet model which we are calling BioVision.

* Add deep trees

* Revert "Add deep trees"

This reverts commit 619f4c0f1b65f68a56bca02f2b398248b7f173db.

* Add paired bootstrap

* Debugs paired bootstrap and adds figures

* Add zero-shot support for biovision

* Add Lu results

* Add predictions for new checkpoints

* Generate latex table with bootstrap auc results

* fix plotting

* idk

* Add predictions for Uu checkpoint

* CI

* Add Experiment Eval Reform

* generic hugging face transformer

* Add some support for a few HuggingFace BERT models

Currently, we only support CXRBERT. This will add support for BlueBERT
and ClinicalBERT. In order for the models to be equivalent, we are
removing the MLP from CXRBERT and using a common simple linear projection
head trained from scratch for all models. The code is not fully functional
yet as we need to also add support a HuggingFace BERT key argument.

* Add HuggingFace BERT key support

The pipeline now supports three different pretrained BERT models from
Huggingface. A handful of bugs were also fixed.

* Add support for mococxr/medaug via BioViL classes

Co-authored-by: Christy Jestin <cjj9000@gmail.com>

* Add pretrained model dir to gitignore

* Add training code

* Add training parsing code

* Update arguments for run train

* Adjust learning rate

* Add flush

* Add wandb support

The pipeline now supports logging and saving of model checkpoints.

* Fix wandb config saving

Currently, the config isn't actually being updated or saved.

* Use custom transforms for BioViL

* Fix image size for biovil to also be 224

* Update gitignore file for vindr-cxr and others

This change adds the vindr-cxr, wandb, and results folder to the
gitignore file so they won't be tracked.

* Add functionality to process the vindr dataset

Before we had no way to process the vindr-cxr dataset since the image
files were all dicom. This change adds in this functionality so the
images can be processed for evaluation.

* Create documentation for the vindr-cxr dataset

This change adds documentation about the vindr-cxr dataset. It includes
info in README files about where to get the data and how to process it.

* Changes to vision projection for CLIP

* Update gitignore file for vindr-cxr and others

This change adds the vindr-cxr, wandb, and results folder to the
gitignore file so they won't be tracked.

* Add functionality to process the vindr dataset

Before we had no way to process the vindr-cxr dataset since the image
files were all dicom. This change adds in this functionality so the
images can be processed for evaluation.

* Create documentation for the vindr-cxr dataset

This change adds documentation about the vindr-cxr dataset. It includes
info in README files about where to get the data and how to process it.

* Fix zero shot for new model types

* Setup evaluation on vindr dataset

This change sets up evaluation capabilities on the vindr dataset. It
also fixes some processing of the vindr dataset.

* Change evaluation for biovision on vindr

Before there had been errors with trying to evaluate a biovision model
on the vindr dataset. This change addresses the errors and the code
runs. However, there might still be a bug with how the image transforms
are done because the auc results are low.

* Improve names and saving for results and bootstrap

This change makes sure zero shot results are saved in a directory for
chexpert or vindr within each model folder. Also, we improve the naming
of pictures from boostrap to include the dataset as well.

* Add previously generated results figures

This change adds some AUC results images that had been created in the
past.

* add new table

* Fix typo with plot paired bootstrap

There was a  typo that was fixed.

* Add code to generate new paired AUC plots

* Adjust table creation code

This change adds in split point support for creating tables for latex.

* Add all final predictions

* Update gitignore to keep final predictions

Makes sures git is keeping tracking of just the final checkpoint
predictions in the predictions folder.

* Update vindr predictions

This change updates the vindr model predictions with the new vindr data
loading infra.

* Create file to preprocess vindr dataset

Before we had been loading and transforming the vindr dataset for every
prediction. This change creates a file to perform the transformations in
advance so the predictions are faster.

* Add ability to run predictions for all models

This change adds a quick way to create predictions for all models at
once. It will skip models that already have predictions in a specific
folder.

* Fix errors with vindr evaluation

Updated the whole vindr evaluation process to fix errors. Did this by
creating new dataset objects. VinDrTestDataset loads and transforms the
vindr data. FastDrTestDataset takes the processed images and loads them
in quickly for evaluation.

* Make small clip adjustments for running training

This changes makes some modifications to help with training different
models.

* Add results figures

This change adds some previously generated results figures to the repo.

Co-authored-by: aakamishra <aakamishra1@gmail.com>
Co-authored-by: greenFantasy <rajatmittal@college.harvard.edu>
Co-authored-by: Christy Jestin <cjj9000@gmail.com>
Co-authored-by: christyjestin <christyjestin@college.harvard.edu>