Cleaning up code and adding documentation (#21)
* Remove outdated hydra support

This change removes the outdatad hydra support since it does not work
with our new model architecture.

* Update documentation for new setup

This change updates the README files with new documentation. This
includes info on the new way to run experiments, zero shot predictions,
and bootstrap results.

* Update bootstrap documentation

This change includes info on paired boostrap docs.